Refactoring modules.
