+++ lib/plugins/capture/capture_test.go
@@ -19,15 +19,16 @@
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-package logger
+package capture
 
 import (
 	"bytes"
+	"log"
 	"net/http"
 	"testing"
 	"time"
 
-	"github.com/xiam/hyperfox/proxy"
+	"github.com/xiam/hyperfox/lib/proxy"
 )
 
 const listenHTTPAddr = `127.0.0.1:37400`
@@ -72,23 +73,31 @@ func TestListenHTTP(t *testing.T) {
 	time.Sleep(time.Millisecond * 100)
 }
 
-func TestLoggerInterface(t *testing.T) {
+func TestWriteCloser(t *testing.T) {
 	var req *http.Request
 	var err error
 
-	// Adding write closer that will receive all the data and then a closing
-	// instruction.
-	px.AddLogger(Stdout{})
+	res := make(chan Response, 10)
+
+	px.AddBodyWriteCloser(New(res))
 
 	urls := []string{
 		"http://golang.org/src/database/sql/",
 		"http://golang.org/",
 		"https://www.example.org",
-		"http://play.golang.org/p/-URiXol0GB",
+		"http://nmap.org",
 	}
 
-	for i := range urls {
+	go func() {
+		for {
+			select {
+			case r := <-res:
+				log.Printf("Captured: %s, %d, (%d bytes) - %dns", r.URL, r.Status, len(r.Body), r.TimeTaken)
+			}
+		}
+	}()
 
+	for i := range urls {
 		// Creating a response writer.
 		wri := newTestResponseWriter()
 
@@ -100,4 +109,6 @@ func TestLoggerInterface(t *testing.T) {
 		// Executing request.
 		px.ServeHTTP(wri, req)
 	}
+
+	time.Sleep(time.Second * 1)
 }
